validated version

+++ exec/clean.c
@@ -6,14 +6,14 @@
 /*   By: ctruchot <ctruchot@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/19 14:00:12 by ctruchot          #+#    #+#             */
-/*   Updated: 2024/04/04 16:16:34 by ctruchot         ###   ########.fr       */
+/*   Updated: 2024/04/17 14:11:17 by ctruchot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "exec.h"
 #include "minishell.h"
 
-int	clean_end(t_exec *exec)
+int	clean_end(t_exec *exec, t_pers *pers)
 {
 	int		j;
 	int		status;
@@ -24,25 +24,12 @@ int	clean_end(t_exec *exec)
 	while (j < exec->total_cmd)
 	{
 		waitpid(exec->pid[j], &status, 0);
-		if (WIFEXITED(status))
-			g_status = WEXITSTATUS(status);
-		else if (WIFSIGNALED(status))
-		{
-			if (status == 139)
-			{
-				ft_putstr_fd("Segmentation fault (core dumped)\n", 2);
-				g_status = status;
-			}
-			else if (status == 131)
-				g_status = status;
-			else
-				g_status = status + 128;
-		}
+		get_status(status, pers);
 		j++;
 		buf = buf->next;
 	}
 	clean_exit_parent(exec, 0);
-	return (g_status);
+	return (pers->status_code);
 }
 
 // to end parent, free all malloc vars and delete the temporary heredoc file
@@ -75,11 +62,12 @@ void	clean_exit_child(t_exec *exec, t_child *child, int err)
 		free_tab_int(exec->fd, exec->total_cmd - 1);
 	if (exec->pid != NULL)
 		free(exec->pid);
-	if (access("/tmp/.tmpheredoc", F_OK) == 0)
-	{
-		if (unlink("/tmp/.tmpheredoc") != 0)
-			ft_putstr_fd(strerror(errno), 2);
-	}
+	if (exec->cmd->type == COMMAND)
+		if (exec->cmd->in)
+			if (exec->cmd->in->mode == DOUBLE)
+				if (access("/tmp/.tmpheredoc", F_OK) == 0)
+					if (unlink("/tmp/.tmpheredoc") != 0)
+						ft_putstr_fd(strerror(errno), 2);
 	if (exec->mini_env != NULL)
 		ft_freetab(exec->mini_env);
 	if (child->current_cmd->code_err == 127)
@@ -93,7 +81,6 @@ void	clean_exit_child(t_exec *exec, t_child *child, int err)
 
 int	clean_exit_fds(t_exec *exec, t_child *child)
 {
-	ft_putstr_fd(strerror(errno), 2);
 	if (child->fdout > 0)
 		close(child->fdout);
 	if (child->fdin > 0)
@@ -117,9 +104,10 @@ void	close_all_fds(t_exec *exec)
 			close(exec->fd[l][1]);
 		l++;
 	}
-	if (access("/tmp/.tmpheredoc", F_OK) == 0)
-	{
-		if (unlink("/tmp/.tmpheredoc") != 0)
-			ft_putstr_fd(strerror(errno), 2);
-	}
+	if (exec->cmd->type == COMMAND)
+		if (exec->cmd->in)
+			if (exec->cmd->in->mode == DOUBLE)
+				if (access("/tmp/.tmpheredoc", F_OK) == 0)
+					if (unlink("/tmp/.tmpheredoc") != 0)
+						ft_putstr_fd(strerror(errno), 2);
 }
